refactor: Move HttpHandler types to same file
