+++ packages/uma/src/routes/Ticket.ts
@@ -8,9 +8,7 @@ import {
   UnsupportedMediaTypeHttpError
 } from '@solid/community-server';
 import { randomUUID } from 'node:crypto';
-import { HttpHandler } from '../util/http/models/HttpHandler';
-import { HttpHandlerContext } from '../util/http/models/HttpHandlerContext';
-import { HttpHandlerResponse } from '../util/http/models/HttpHandlerResponse';
+import { HttpHandler, HttpHandlerContext, HttpHandlerResponse } from '../util/http/models/HttpHandler';
 import { verifyRequest } from '../util/HttpMessageSignatures';
 import { array, reType } from '../util/ReType';
 import { Permission } from '../views/Permission';
