+++ packages/uma/src/util/HttpMessageSignatures.ts
@@ -2,15 +2,15 @@ import { UnauthorizedHttpError, type AlgJwk, BadRequestHttpError, InternalServer
 import { httpbis, type SigningKey, type Request as SignRequest } from 'http-message-signatures';
 import { verifyMessage } from 'http-message-signatures/lib/httpbis';
 import { type SignatureParameters, type VerifierFinder, type VerifyingKey } from 'http-message-signatures/lib/types';
-import type { HttpHandlerRequest } from './http/models/HttpHandlerRequest';
+import { HttpHandlerRequest } from './http/models/HttpHandler';
 import buildGetJwks from 'get-jwks';
 import crypto from 'node:crypto';
 
 const authParserMod = import('@httpland/authorization-parser');
 
 export async function signRequest(
-  url: string, 
-  request: RequestInit & Omit<SignRequest, 'url'>, 
+  url: string,
+  request: RequestInit & Omit<SignRequest, 'url'>,
   jwk: AlgJwk
 ): Promise<RequestInit & SignRequest> {
   const key: SigningKey = {
@@ -52,10 +52,10 @@ export async function verifyRequest(
   signer?: string,
 ): Promise<boolean> {
   signer = signer ?? await extractRequestSigner(request);
-  
+
   if (signer.startsWith('"')) signer = signer.slice(1);
   if (signer.endsWith('"')) signer = signer.slice(0,-1);
-  
+
   const jwks = buildGetJwks();
 
   const keyLookup: VerifierFinder = async (params: SignatureParameters) => {
@@ -67,7 +67,7 @@ export async function verifyRequest(
         alg: alg ?? '',
         kid: keyid ?? '',
       })
-    
+
       if (!alg) throw new BadRequestHttpError('Invalid HTTP message Signature parameters.');
       // if (alg === 'EdDSA') throw new InternalServerError('EdDSA signing is not supported');
       // if (alg === 'ES256K') throw new InternalServerError('ES256K signing is not supported');
@@ -83,7 +83,7 @@ export async function verifyRequest(
           } catch (err) { console.log(err); return null }
         },
       };
-      
+
       return verifier;
 
     } catch (err) {
@@ -111,4 +111,3 @@ const algMap: Record<string, AlgParams> = {
   'RS384': { name: 'RSASSA-PKCS1-v1_5', hash: 'SHA-384' },
   'RS512': { name: 'RSASSA-PKCS1-v1_5', hash: 'SHA-512' },
 }
-
