+++ packages/uma/src/util/http/models/HttpHandler.ts
@@ -1,7 +1,23 @@
 import { AsyncHandler } from '@solid/community-server';
-import { HttpHandlerResponse } from './HttpHandlerResponse';
-import { HttpHandlerContext } from './HttpHandlerContext';
+import { OutgoingHttpHeaders } from 'http';
 
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface HttpHandlerContext {
+  request: HttpHandlerRequest;
+}
+
+export interface HttpHandlerRequest<B = unknown> {
+  url: URL;
+  method: string;
+  parameters?: { [key: string]: string };
+  headers: { [key: string]: string };
+  body?: B;
+}
+
+export interface HttpHandlerResponse<B = unknown> {
+  body?: B;
+  headers?: OutgoingHttpHeaders;
+  status: number;
+}
+
 export abstract class HttpHandler<C extends HttpHandlerContext = HttpHandlerContext>
   extends AsyncHandler<C, HttpHandlerResponse> { }
