--- packages/uma/src/util/http/models/HttpHandlerContext.ts
@@ -1,6 +1,0 @@
-import { HttpHandlerRequest } from './HttpHandlerRequest';
-import { HttpHandlerRoute } from './HttpHandlerRoute';
-
-export interface HttpHandlerContext {
-  request: HttpHandlerRequest;
-}
