--- packages/uma/src/util/http/models/HttpHandlerController.ts
@@ -1,10 +1,0 @@
-import { HttpHandlerContext } from './HttpHandlerContext';
-import { HttpHandlerRoute } from './HttpHandlerRoute';
-
-export class HttpHandlerController<C extends HttpHandlerContext = HttpHandlerContext> {
-
-  constructor(
-    public routes: HttpHandlerRoute<C>[],
-  ) { }
-
-}
