--- packages/uma/src/util/http/models/HttpHandlerRequest.ts
@@ -1,7 +1,0 @@
-export interface HttpHandlerRequest<B = unknown> {
-  url: URL;
-  method: string;
-  parameters?: { [key: string]: string };
-  headers: { [key: string]: string };
-  body?: B;
-}
