--- packages/uma/src/util/http/models/HttpHandlerResponse.ts
@@ -1,7 +1,0 @@
-import { OutgoingHttpHeaders } from 'http';
-
-export interface HttpHandlerResponse<B = unknown> {
-  body?: B;
-  headers?: OutgoingHttpHeaders;
-  status: number;
-}
